+++ workflows/SUEP_utils.py
@@ -1,11 +1,9 @@
 import numpy as np
 import pandas as pd
 import awkward as ak
+import fastjet
 import vector
 vector.register_awkward()
-
-from workflows.math_utils import *
-from workflows.inference_utils import *
 
 def ClusterMethod(self, indices, tracks, 
                   SUEP_cand, ISR_cand, 
@@ -272,116 +270,72 @@ def ConeMethod(self, indices, tracks,
                 self.out_vars.loc[indices, "ISR_phi_CO"] = ISR_cand.phi
                 self.out_vars.loc[indices, "ISR_mass_CO"] = ISR_cand.mass
 
-def SSDMethod(self, indices, events, out_label=''):
-    #####################################################################################
-    # ---- ML Analysis
-    # Each event is converted into an input for the ML models. Using ONNX, we run
-    # inference on each event to obtain a prediction of the class (SUEP or QCD).
-    # N.B.: Conversion is done elsewhere.
-    # N.B.: The inference skips the lost tracks for now.
-    #####################################################################################
+def sphericity(particles, r):
+    norm = ak.sum(particles.p ** r, axis=1, keepdims=True)
+    s = np.array([[
+                   ak.sum(particles.px * particles.px * particles.p ** (r-2.0), axis=1 ,keepdims=True)/norm,
+                   ak.sum(particles.px * particles.py * particles.p ** (r-2.0), axis=1 ,keepdims=True)/norm,
+                   ak.sum(particles.px * particles.pz * particles.p ** (r-2.0), axis=1 ,keepdims=True)/norm
+                  ],
+                  [
+                   ak.sum(particles.py * particles.px * particles.p ** (r-2.0), axis=1 ,keepdims=True)/norm,
+                   ak.sum(particles.py * particles.py * particles.p ** (r-2.0), axis=1 ,keepdims=True)/norm,
+                   ak.sum(particles.py * particles.pz * particles.p ** (r-2.0), axis=1 ,keepdims=True)/norm
+                  ],
+                  [
+                   ak.sum(particles.pz * particles.px * particles.p ** (r-2.0), axis=1 ,keepdims=True)/norm,
+                   ak.sum(particles.pz * particles.py * particles.p ** (r-2.0), axis=1 ,keepdims=True)/norm,
+                   ak.sum(particles.pz * particles.pz * particles.p ** (r-2.0), axis=1 ,keepdims=True)/norm
+                  ]])
+    s = np.squeeze(np.moveaxis(s, 2, 0),axis=3)
+    evals = np.sort(np.linalg.eigvalsh(s))
+    return evals
 
-    if self.do_inf:    
+def rho(number, jet, tracks, deltaR, dr=0.05):
+    r_start = number*dr
+    r_end = (number+1)*dr
+    ring = (deltaR > r_start) & (deltaR < r_end)
+    rho_values = ak.sum(tracks[ring].pt, axis=1)/(dr*jet.pt)
+    return rho_values
 
-        pred_dict = {}
-        ort_infs = {}
-        options = ort.SessionOptions() 
-        options.inter_op_num_threads = 1 # number of threads used to parallelize the execution of the graph (across nodes). Default is 0 to let onnxruntime choose.
-        for model in self.ssd_models:
-              ort_infs.update({model: ort.InferenceSession('data/onnx_models/resnet_{}_{}.onnx'.format(model,self.era))})
-        # In order to avoid memory issues convert events to images and run inference in batches
-        # also exploits the numba-compiled convert_to_images function
-        for i in range(0, len(inf_cands), self.batch_size):
-            if i + self.batch_size > len(inf_cands): self.batch_size = len(inf_cands) - i
-            batch = inf_cands[i:i+self.batch_size]
-            imgs = convert_to_images(self, batch)
-            for model in self.ssd_models:
-                batch_resnet_jets = run_inference(self, imgs, ort_infs[model])
-                if i == 0: resnet_jets = batch_resnet_jets
-                else: resnet_jets = np.concatenate((resnet_jets, batch_resnet_jets))    
-                pred_dict.update({model: resnet_jets[:,1]}) #highest SUEP prediction per event
-
-        for model in self.ssd_models:
-            self.out_vars.loc[indices, model+"_ssd"+out_label] = pred_dict[model]
-    else:
-        for c in self.ssd_models: self.out_vars.loc[indices, c+"_ssd"+out_label] = np.nan
-
-def DGNNMethod(self, indices, events, SUEP_tracks, SUEP_cand, out_label=''):
-    #####################################################################################
-    # ---- ML Analysis
-    # Each event is converted into an input for the ML models. Using ONNX, we run
-    # inference on each event to obtain a prediction of the class (SUEP or QCD).
-    # N.B.: Conversion is done elsewhere.
-    #####################################################################################
-
-    if self.do_inf:    
+def FastJetReclustering(self, tracks, r, minPt):
         
-        import yaml
-        import torch
-        from torch import nn
-        from torch_geometric.data import Data
+    jetdef = fastjet.JetDefinition(fastjet.antikt_algorithm, r)        
+    cluster = fastjet.ClusterSequence(tracks, jetdef)
 
-        from workflows.SUEPNet import Net        
+    # have to set min_pt = 0 and cut later to avoid some memory issues
+    # FIXME: should try to understand this failure
+    ak_inclusive_jets = cluster.inclusive_jets()[:] 
+    ak_inclusive_cluster = cluster.constituents()[:]
 
-        device = torch.device('cpu')
+    # apply minimum pT cut
+    minPtCut = ak_inclusive_jets.pt > minPt
+    ak_inclusive_jets = ak_inclusive_jets[minPtCut]
+    ak_inclusive_cluster = ak_inclusive_cluster[minPtCut]
 
-        # consistency check
-        assert len(self.dgnn_model_names) == len(self.configs)
+    return ak_inclusive_jets, ak_inclusive_cluster
         
-        for model_name, config in zip(self.dgnn_model_names, self.configs):
-            
-            model_path = 'data/' + model_name + '.pt'
-            # initialize model with original configurations and import the weights
-            config = yaml.safe_load(open(config))
-            suep = Net(out_dim=config['model_pref']['out_dim'], 
-                    hidden_dim=config['model_pref']['hidden_dim']).to(device)
-            suep.load_state_dict(torch.load(model_path, map_location=torch.device('cpu'))['model'])
-            suep = suep.float()
-            suep.eval()
-            sigmoid = torch.nn.Sigmoid()
-            
-            results = np.array([])
-            for i in range(0, len(events), self.batch_size):
-                
-                # define batch and convert objects in a coordinate frame
-                batch = events[i:i+self.batch_size]
-                this_batch_size = len(batch)
-                batch_SUEP_cand = SUEP_cand[i:i+self.batch_size]
-                batch = GNN_convertEvents(self, batch, batch_SUEP_cand)
-                
-                with torch.no_grad():
-                    
-                    # count how many tracks per event are nonzero: dimensions of (events, tracks)
-                    Nlc = np.count_nonzero(batch[:,:,0], axis=1)
-                    
-                    # reshape the batch for correct DGNN format
-                    x_pf = None  # dims: (events times tracks, 4)
-                    x_pf_batch = None   #dim: (events times tracks)
-                    for idx in range(this_batch_size):
-                        # flatten the batch
-                        if idx == 0: x_pf = batch[idx, :Nlc[idx], :]
-                        else: x_pf = np.vstack((x_pf, batch[idx, :Nlc[idx], :]))
-                        # event indices
-                        if idx == 0: x_pf_batch = np.ones(Nlc[idx])*idx
-                        else: x_pf_batch = np.concatenate((x_pf_batch, np.ones(Nlc[idx])*idx))
+def getTopTwoJets(self, tracks, indices, ak_inclusive_jets, ak_inclusive_cluster):
+    # order the reclustered jets by pT (will take top 2 for ISR removal method)
+    highpt_jet = ak.argsort(ak_inclusive_jets.pt, axis=1, ascending=False, stable=True)
+    jets_pTsorted = ak_inclusive_jets[highpt_jet]
+    clusters_pTsorted = ak_inclusive_cluster[highpt_jet]     
 
-                    # convert to torch
-                    x_pf = torch.from_numpy(x_pf).float()
-                    x_pf_batch = torch.from_numpy(x_pf_batch).float()
+    # at least 2 tracks in highest pt jet
+    highpt_cands = clusters_pTsorted[:,0]                    # tracks for highest pt jet         
+    singletrackCut = (ak.num(highpt_cands)>1)             
+    jets_pTsorted = jets_pTsorted[singletrackCut]          
+    clusters_pTsorted = clusters_pTsorted[singletrackCut]
+    tracks = tracks[singletrackCut]
+    indices = indices[singletrackCut]
 
-                    # batch predictions
-                    out = suep(x_pf, x_pf_batch)
+    # number of constituents per jet, sorted by pT
+    nconst_pTsorted = ak.num(clusters_pTsorted, axis=-1)
 
-                    # normalize the outputs
-                    nn1 = out[0][:,0]
-                    nn1 = sigmoid(nn1)
-                    results = np.concatenate((results, nn1.cpu().numpy()))
-            
-            self.out_vars.loc[indices, model_name + "_GNN"+out_label] = results
-            
-            eigs = sphericity(SUEP_tracks, 1.0) #Set r=1.0 for IRC safe
-            self.out_vars.loc[indices, "SUEP_nconst_GNN"+out_label] = ak.num(SUEP_tracks)
-            self.out_vars.loc[indices, "SUEP_S1_GNN"+out_label] = 1.5 * (eigs[:,1]+eigs[:,0])
+    # Top 2 pT jets. If jet1 has fewer tracks than jet2 then swap
+    SUEP_cand = ak.where(nconst_pTsorted[:,1]<=nconst_pTsorted[:,0],jets_pTsorted[:,0],jets_pTsorted[:,1])
+    ISR_cand = ak.where(nconst_pTsorted[:,1]>nconst_pTsorted[:,0],jets_pTsorted[:,0],jets_pTsorted[:,1])
+    SUEP_cluster_tracks = ak.where(nconst_pTsorted[:,1]<=nconst_pTsorted[:,0], clusters_pTsorted[:,0], clusters_pTsorted[:,1])
+    ISR_cluster_tracks = ak.where(nconst_pTsorted[:,1]>nconst_pTsorted[:,0], clusters_pTsorted[:,0], clusters_pTsorted[:,1])
 
-    else:
-        for c in self.dgnn_model_names: self.out_vars.loc[indices, c+"_GNN"+out_label] = np.nan
+    return tracks, indices, (SUEP_cand, ISR_cand, SUEP_cluster_tracks, ISR_cluster_tracks)
